pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "davies-snow-sensor"
version = "0.1.0"
description = "Davies Snow Sensor software stack"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "RPi.GPIO>=0.7.1",
    "gpiozero>=2.0",
    "adafruit-blinka>=8.0.0",
    "adafruit-circuitpython-rfm9x>=2.0.0",
    "PyYAML>=6.0",
    "w1thermsensor>=2.3.0",
    "polars>=0.20.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
]

[project.optional-dependencies]
analysis = [
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]
notebook = [
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
]
test = [
    "pytest>=7.0.0",
]
dev = [
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
    "pytest>=7.0.0",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
